Restructure repo for "go get".
